move concesion helper into business rules lib
