+++ packages/gafl-webapp-service/src/pages/concessions/disability/update-transaction.js
@@ -1,5 +1,5 @@
 import { DISABILITY_CONCESSION } from '../../../uri.js'
-import * as concessionHelper from '../../../processors/concession-helper.js'
+import { addDisabled, removeDisabled } from '@defra-fish/business-rules-lib'
 import { CONCESSION_PROOF } from '../../../processors/mapping-constants.js'
 export const disabilityConcessionTypes = {
   pipDla: 'pip-dla',
@@ -16,13 +16,13 @@ export default async request => {
   const permission = await request.cache().helpers.transaction.getCurrentPermission()
 
   if (payload['disability-concession'] === disabilityConcessionTypes.pipDla) {
-    concessionHelper.addDisabled(permission, CONCESSION_PROOF.NI, payload['ni-number'])
+    addDisabled(permission, CONCESSION_PROOF.NI, payload['ni-number'])
     Object.assign(permission, { licenceLength: '12M', licenceStartTime: '0' })
   } else if (payload['disability-concession'] === disabilityConcessionTypes.blueBadge) {
-    concessionHelper.addDisabled(permission, CONCESSION_PROOF.blueBadge, payload['blue-badge-number'])
+    addDisabled(permission, CONCESSION_PROOF.blueBadge, payload['blue-badge-number'])
     Object.assign(permission, { licenceLength: '12M', licenceStartTime: '0' })
   } else {
-    concessionHelper.removeDisabled(permission)
+    removeDisabled(permission)
   }
 
   await request.cache().helpers.transaction.setCurrentPermission(permission)
