+++ packages/gafl-webapp-service/src/pages/contact/contact/route.js
@@ -2,9 +2,8 @@ import { CONTACT, LICENCE_LENGTH, DATE_OF_BIRTH, LICENCE_TO_START } from '../../
 import pageRoute from '../../../routes/page-route.js'
 import GetDataRedirect from '../../../handlers/get-data-redirect.js'
 import Joi from 'joi'
-import { validation } from '@defra-fish/business-rules-lib'
+import { hasJunior, validation } from '@defra-fish/business-rules-lib'
 import { isPhysical } from '../../../processors/licence-type-display.js'
-import { hasJunior } from '../../../processors/concession-helper.js'
 import { nextPage } from '../../../routes/next-page.js'
 import { mobilePhoneValidator } from '../../../processors/contact-validator.js'
 import { HOW_CONTACTED } from '../../../processors/mapping-constants.js'
