+++ packages/gafl-webapp-service/src/processors/renewals-write-cache.js
@@ -1,9 +1,8 @@
 import moment from 'moment-timezone'
 import db from 'debug'
 import { LICENCE_TYPE, NAME, ADDRESS_LOOKUP, CONTACT, LICENCE_FULFILMENT, LICENCE_CONFIRMATION_METHOD } from '../uri.js'
-import { SERVICE_LOCAL_TIME } from '@defra-fish/business-rules-lib'
+import { addDisabled, ageConcessionHelper, SERVICE_LOCAL_TIME } from '@defra-fish/business-rules-lib'
 import * as constants from './mapping-constants.js'
-import { ageConcessionHelper, addDisabled } from './concession-helper.js'
 import { licenceToStart } from '../pages/licence-details/licence-to-start/update-transaction.js'
 import { licenseTypes } from '../pages/licence-details/licence-type/route.js'
 import { salesApi } from '@defra-fish/connectors-lib'
